update Quantity and Margin on set new Category
Only the first option is setting the filter to the other one, either is Quantity or Margin
So, in case you choose Margin first, the Margin selector is not going to be filter by what's inside Quantity.
if you choose Quantity filter, Quantity is not going to be filtered by what's selected in Margin!